refactor: use internal instead pkg
