+++ internal/changelog/markdown.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/haunt98/changeloguru/pkg/convention"
-	"github.com/haunt98/changeloguru/pkg/markdown"
+	"github.com/haunt98/changeloguru/internal/convention"
+	"github.com/haunt98/changeloguru/internal/markdown"
 	"github.com/haunt98/clock"
 )
