+++ internal/changelog/testdata/TestGenerateMarkdown/empty_old_data.golden
