+++ internal/cli/action_generate.go
@@ -7,10 +7,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/haunt98/changeloguru/pkg/changelog"
-	"github.com/haunt98/changeloguru/pkg/convention"
-	"github.com/haunt98/changeloguru/pkg/git"
-	"github.com/haunt98/changeloguru/pkg/markdown"
+	"github.com/haunt98/changeloguru/internal/changelog"
+	"github.com/haunt98/changeloguru/internal/convention"
+	"github.com/haunt98/changeloguru/internal/git"
+	"github.com/haunt98/changeloguru/internal/markdown"
 	"github.com/pkg/diff"
 	"github.com/pkg/diff/write"
 	"github.com/urfave/cli/v2"
