+++ internal/convention/commit_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/haunt98/changeloguru/pkg/git"
+	"github.com/haunt98/changeloguru/internal/git"
 	"github.com/sebdah/goldie/v2"
 	"github.com/stretchr/testify/assert"
 )
