+++ internal/convention/testdata/TestNewCommit/Commit_message_with_no_body.golden
