+++ internal/convention/testdata/TestNewCommit/Commit_message_with_scope.golden
