+++ internal/convention/testdata/TestNewCommit/Uppercase.golden
