+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/haunt98/changeloguru/pkg/cli"
+import "github.com/haunt98/changeloguru/internal/cli"
 
 func main() {
 	app := cli.NewApp()
